split buiseness logic from html on the error page
